feat: implementing generation of dependencies json
