+++ packages/tools/pluggable-widgets-tools/configs/rollup.config.js
@@ -1,5 +1,5 @@
-import { existsSync, mkdirSync, promises } from "fs";
-import { basename, dirname, extname, join, relative } from "path";
+import { existsSync, mkdirSync } from "fs";
+import { join, relative } from "path";
 import alias from "@rollup/plugin-alias";
 import { getBabelInputPlugin, getBabelOutputPlugin } from "@rollup/plugin-babel";
 import commonjs from "@rollup/plugin-commonjs";
@@ -9,7 +9,6 @@ import replace from "@rollup/plugin-replace";
 import typescript from "@rollup/plugin-typescript";
 import url from "@rollup/plugin-url";
 import { red, yellow } from "colors";
-import { copy, readJson } from "fs-extra";
 import postcss from "postcss";
 import postcssUrl from "postcss-url";
 import loadConfigFile from "rollup/dist/loadConfigFile";
@@ -21,6 +20,7 @@ import { terser } from "rollup-plugin-terser";
 import { cp } from "shelljs";
 import { zip } from "zip-a-folder";
 import { widgetTyping } from "./rollup-plugin-widget-typing";
+import { collectNativeDependencies } from "./rollup-plugin-collect-native-deps";
 import {
     editorConfigEntry,
     isTypescript,
@@ -94,7 +94,15 @@ export default async args => {
                     }),
                     ...(i === 0 ? getClientComponentPlugins() : []),
                     json(),
-                    copyNativeDependency({ nodeModulesPath: join(outDir, "node_modules") }),
+                    ...(i === 0
+                        ? [
+                              collectNativeDependencies({
+                                  outputDir: outDir,
+                                  externals: nativeExternal,
+                                  widgetName
+                              })
+                          ]
+                        : []),
                     ...getCommonPlugins({
                         sourceMaps: false,
                         extensions: [`.${os}.js`, ".native.js", ".js", ".jsx", ".ts", ".tsx"],
@@ -283,94 +291,3 @@ const nativeExternal = [
     "react-native-vector-icons",
     "react-navigation"
 ];
-
-function copyNativeDependency({ nodeModulesPath }) {
-    // 1. Identify whether a dependency has a native module
-    // 2. Copy native modules into node_modules folder at com folder level
-    const nativeDependencies = []; // todo: fix watch mode
-    return {
-        name: "copy-react-native-modules",
-        async resolveId(source) {
-            if (source.startsWith(".")) {
-                return null;
-            }
-
-            try {
-                const packageDir = dirname(require.resolve(`${source}/package.json`));
-
-                if (await hasNativeCode(packageDir)) {
-                    if (!nativeDependencies.some(x => x.name === source)) {
-                        nativeDependencies.push({ name: source, dir: packageDir });
-                    }
-                    return { id: source, external: true };
-                }
-                return null;
-            } catch (e) {
-                return null;
-            }
-        },
-        async writeBundle() {
-            await Promise.all(
-                nativeDependencies.map(async dependency => {
-                    await copyJsModule(dependency.dir, join(nodeModulesPath, dependency.name));
-                    for (const transitiveDependency of await getTransitiveDependencies(dependency.name)) {
-                        await copyJsModule(
-                            dirname(require.resolve(`${transitiveDependency}/package.json`)),
-                            join(nodeModulesPath, dependency.name, "node_modules", transitiveDependency)
-                        );
-                    }
-                })
-            );
-            // const nativeDependencies = Object.values(bundle)
-            //     .flatMap(c => c.imports.concat(c.dynamicImports))
-            //     .filter(d => d.startsWith("react-native-"))
-            //     .map(d => d.split("/")[0]);
-            //
-            // const packagedToCopy = withTransitiveDependencies(nativeDependencies);
-            // withTransitiveDependencies(MODULES_SHIPPED_WITH_CLIENT, true).forEach(d => packagedToCopy.delete(d));
-            //
-            // await copyPackages(packagedToCopy, dest);
-        }
-    };
-}
-
-async function hasNativeCode(dir) {
-    const packageContent = await promises.readdir(dir, { withFileTypes: true });
-    return (
-        packageContent.some(file => /^(ios|android|.*\.podspec)$/i.test(file.name)) ||
-        packageContent.some(file => file.isDirectory() && hasNativeCode(join(dir, file.name)))
-    );
-}
-
-async function getTransitiveDependencies(packageName) {
-    const queue = [packageName];
-    const result = new Set();
-    while (queue.length) {
-        const next = queue.shift();
-        if (result.has(next)) {
-            continue;
-        }
-        const isExternal = nativeExternal.some(external =>
-            external instanceof RegExp ? external.test(next) : external === next
-        );
-        if (isExternal) {
-            continue;
-        }
-
-        if (next !== packageName) {
-            result.add(next);
-        }
-        const packageJson = await readJson(require.resolve(`${next}/package.json`));
-        queue.push(...Object.keys(packageJson.dependencies ?? {}));
-    }
-    return Array.from(result);
-}
-
-async function copyJsModule(from, to) {
-    await copy(from, to, {
-        filter: async path =>
-            (await promises.lstat(path)).isDirectory()
-                ? !["android", "ios", ".github", "__tests__"].includes(basename(path))
-                : [".js", ".jsx", ".json"].includes(extname(path)) || basename(path).toLowerCase().includes("license")
-    });
-}
